Consolidate task queue validation
Consolidated task queue validation. Also added more stringent validation to guard against leading and trailing whitespace.
